vertigo/functions/meeting-processor/requirements.txt
google-cloud-firestore
google-auth
google-generativeai==0.8.5
functions-framework
langfuse 